cluster/pyproject.toml
[project]
name = "cluster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
]

[tool.ruff]
line-length = 80

[tool.ruff.format]
quote-style = "single"
